requirements.txt
django
Pillow
django-bootstrap-form
pydub
ffmpeg-python
djangorestframework